Add tasking codegen, parser improvements, and type system enhancements
Parser:
- Add parse_dotted_name() and name_to_string() helper methods
- Support character literals in enumeration types
- Add range constraint parsing for derived types

Type system:
- Universal_Integer now recognized as discrete type
- Universal_Integer/Universal_Real recognized as numeric types
- Add Wide_Character and Wide_String predefined types

Lowering/codegen:
- Add task type and task body lowering support
- Improve SubtypeIndication vs ArrayTypeDef handling
- Handle SelectedName in parameter type marks

Tests:
- Add task body compilation tests
- Add task with entry compilation tests
- Add delay statement compilation tests

+++ tests/test_ir.py
@@ -38,7 +38,7 @@ def test_ir_type_from_bits():
     assert ir_type_from_bits(1) == IRType.BYTE
     assert ir_type_from_bits(8) == IRType.BYTE
     assert ir_type_from_bits(16) == IRType.WORD
-    assert ir_type_from_bits(32) == IRType.WORD
+    assert ir_type_from_bits(32) == IRType.DWORD  # 32-bit values use DWORD
 
 
 # ============================================================================
